disabled blog link and added more contents on frontend

+++ backend/app/Http/Controllers/AuthController.php
@@ -6,16 +6,26 @@ use Illuminate\Http\Request;
 use App\Models\User;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Validation\ValidationException;
-
+use Illuminate\Support\Facades\Validator;
+ 
 class AuthController extends Controller
 {
     public function register(Request $request)
     {
-        $request->validate([
-            'name' => 'required|string|max:255',
-            'email' => 'required|string|email|unique:users',
-            'password' => 'required|string|min:6'
-        ]);
+        try{
+
+       
+       $validator = Validator::make($request->all(), [
+        "email"=> "email|required|unique:users",
+        "password"=> "required",
+        "name"=>"required"
+       ]);
+
+       if ($validator->fails()){
+        return response()->json(['success'=> false,'errors'=> $validator->errors()], 401);
+       }
+
+
 
         $user = User::create([
             'name' => $request->name,
@@ -23,40 +33,62 @@ class AuthController extends Controller
             'password' => Hash::make($request->password),
         ]);
 
-        return response()->json(['message' => 'User registered successfully']);
+        return response()->json(['success'=>'true', 'message'=> 'User successfully registered'], 200);
+       
+    }catch(\Exception $e){
+        return response()->json(['success'=>false, 'message'=>$e->getMessage()], 500);
+    }
+
+
+        
     }
 
     public function login(Request $request)
     {
-        // return response()->json($request->all());
-        $request->validate([
-            'email' => 'required|email',
-            'password' => 'required'
-        ]);
 
-        $user = User::where('email', $request->email)->first();
-        
-        if (!$user || !Hash::check($request->password, $user->password)) {
-            return response()->json(['message' => 'Invalid Credentials'], 401);
+        try{
+
+            $validator = Validator::make($request->all(), [
+'email' => 'required|email',
+                'password' => 'required'
+            ]);
+            if($validator->fails()){
+                return response()->json(['success'=> false,'errors'=> $validator->errors()],401);
+            }
+    
+            $user = User::where('email', $request->email)->first();
+            
+            if (!$user || !Hash::check($request->password, $user->password)) {
+                return response()->json(['success'=>false,'message' => 'Invalid Credentials'], 401);
+    
+            }
+    
+            $user->tokens()->delete();
+    
+            return response()->json(['success'=>true,'message'=> 'Login Successful',
+                'token' => $user->createToken('authToken')->plainTextToken,
+                'user' => $user
+            ], 200);
+
+        }catch(\Exception $e){
+            return response()->json(['success'=>false, 'message'=>$e->getMessage()], 500);
 
         }
-
-        $user->tokens()->delete();
-
-        return response()->json([
-            'token' => $user->createToken('authToken')->plainTextToken,
-            'user' => $user
-        ]);
+        
+       
     }
 
     public function logout(Request $request)
     {
+        try{
+
+       
         $request->user()->tokens()->delete();
-        return response()->json(['message' => 'Logged out successfully']);
+        return response()->json(['success'=>true,'message'=> 'Logout Successful'], 200);
+    }catch(\Exception $e){
+        return response()->json(['success'=>false,'message'=>$e->getMessage()], 500);
+    }
     }
 
-    public function user(Request $request)
-    {
-        return response()->json($request->user());
-    }
+  
 }
